feat [#94]: Cases App
- Carves out new "cases" app from "poc"
- Moves models into the app. Migrations' state and db operations separated.
- Refactored of references of models, serializers, views and urls.
- Fixed unit tests. All unit tests passing.
