+++ cases/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CasesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'cases'
