+++ cases/tests/factories.py
@@ -1,0 +1,11 @@
+import factory
+
+
+class CaseFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = "cases.Case"
+
+
+class CaseLitigantFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = "cases.CaseLitigant"
