+++ poc/api/urls.py
@@ -1,20 +1,14 @@
 from django.urls import path
 
 from poc.api.views import (
-    ListCreateCaseAPI,
     ListCreateMessageAPI,
     ListCreateThreadAPI,
     ListCreateUploadedFileAPI,
-    RetrieveUpdateCaseAPI,
     RetrieveUpdateDestroyUploadedFileAPI,
 )
 
 app_name = "poc"
 urlpatterns = [
-    path("cases/", ListCreateCaseAPI.as_view(), name="cases"),
-    path(
-        "cases/<uuid:case_uuid>/", RetrieveUpdateCaseAPI.as_view(), name="case_detail"
-    ),
     path(
         "cases/<uuid:case_uuid>/exhibits/",
         ListCreateUploadedFileAPI.as_view(),
