+++ poc/api/views.py
@@ -1,8 +1,6 @@
-from django.db.models import Q
 from django.shortcuts import get_object_or_404
 from rest_framework.generics import (
     ListCreateAPIView,
-    RetrieveUpdateAPIView,
     RetrieveUpdateDestroyAPIView,
 )
 from rest_framework.pagination import PageNumberPagination
@@ -10,68 +8,21 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from cases.models import Case
 from poc.api.serializers import (
-    CaseCompactSerializer,
-    CaseSerializer,
     ChatMessageSerializer,
     ChatThreadSerializer,
     UploadedFileSerializer,
 )
 from poc.langchain.chat_agent import send_message
-from poc.models import Case, ChatMessage, ChatThread, UploadedFile
+from poc.models import ChatMessage, ChatThread, UploadedFile
 
 __all__ = [
-    "ListCreateCaseAPI",
-    "RetrieveUpdateCaseAPI",
     "ListCreateUploadedFileAPI",
     "RetrieveUpdateDestroyUploadedFileAPI",
     "ListCreateThreadAPI",
     "ListCreateMessageAPI",
 ]
-
-
-class ListCreateCaseAPI(ListCreateAPIView):
-    permission_classes = [IsAuthenticated]
-    serializer_class = CaseSerializer
-
-    def get_queryset(self):
-        queryset = Case.objects.all().order_by("-id")
-
-        # check for query param 'search'
-        search = self.request.query_params.get("search")
-        if search:
-            if len(search.strip()) > 2:
-                queryset = queryset.filter(
-                    Q(title__icontains=search) | Q(case_number__icontains=search)
-                )
-            else:
-                queryset = queryset.none()
-
-        return queryset
-
-
-class RetrieveUpdateCaseAPI(RetrieveUpdateAPIView):
-    permission_classes = [IsAuthenticated]
-    lookup_field = "uuid"
-    lookup_url_kwarg = "case_uuid"
-    http_method_names = ["get", "patch", "options"]
-
-    def get_queryset(self):
-        if self.request.query_params.get("compact") == "true":
-            return Case.objects.all()
-
-        return Case.objects.prefetch_related(
-            "case_litigants__litigant",
-            "case_litigants__role",
-        )
-
-    def get_serializer_class(self):
-        # check for query param 'compact'
-        if self.request.query_params.get("compact") == "true":
-            return CaseCompactSerializer
-
-        # otherwise return full serializer
-        return CaseSerializer
 
 
 class ListCreateUploadedFileAPI(ListCreateAPIView):
