+++ poc/tests/factories.py
@@ -1,16 +1,6 @@
 import factory
 
 
-class CaseFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = "poc.Case"
-
-
-class CaseLitigantFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = "poc.CaseLitigant"
-
-
 class ChatThreadFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = "poc.ChatThread"
